pyproject.toml
[project]
name = "hiring-test-belg"
version = "0.1.0"
description = "Invoice Generator from Image using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.28.0",
    "litellm>=1.0.0",
    "reportlab>=4.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]
